Remove duplication by sharing `DelegateSpy`.
