+++ QuizCore/QuizCoreTests/QuizTests.swift
@@ -46,22 +46,5 @@ class QuizTests: XCTestCase {
         assertEqual(delegate.completedQuizzes[1], [("Q1", "A1-1"), ("Q2", "A2-2")])
     }
     
-    // MARK:- Helpers
-    
-    private class DelegateSpy: QuizDelegate {
-        
-        var completedQuizzes: [[(String, String)]] = []
-        var answerCompletions: [(String) -> Void] = []
-        
-        func answer(for question: String, completion: @escaping (String) -> Void) {
-            self.answerCompletions.append(completion)
-        }
-        
-        func didCompleteQuiz(with answers: [(question: String, answer: String)]) {
-            completedQuizzes.append(answers)
-        }
-        
-    }
-    
 }
